refactor: extract SafeLocalStorage
