+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './error.utils';
 export * from './sanitize.utils';
 export * from './safe-parsers.utils';
 export * from './validations';
+export * from './safe-local-storage';
