+++ src/utils/safe-local-storage.ts
@@ -1,0 +1,170 @@
+export interface StorageManager {
+  set(key: string, value: unknown): boolean;
+  get(key: string): unknown;
+  remove(key: string): boolean;
+  isAvailable(): boolean;
+  clear(): void;
+  getSize(): number;
+}
+
+export class SafeLocalStorage implements StorageManager {
+  private readonly available: boolean;
+  private readonly memoryFallback: Map<string, string> = new Map();
+
+  constructor() {
+    this.available = this.checkAvailability();
+  }
+
+  private checkAvailability(): boolean {
+    try {
+      const testKey = '__tracelog_test__';
+      window.localStorage.setItem(testKey, 'test');
+      window.localStorage.removeItem(testKey);
+      return true;
+    } catch {
+      console.warn('[TraceLog] localStorage not available, using memory fallback');
+      return false;
+    }
+  }
+
+  set(key: string, value: unknown): boolean {
+    try {
+      const serialized = JSON.stringify(value);
+
+      if (this.available) {
+        const estimatedSize = serialized.length + key.length;
+
+        if (estimatedSize > 1024 * 1024) {
+          console.warn('[TraceLog] Data too large for localStorage, using memory fallback');
+          this.memoryFallback.set(key, serialized);
+          return true;
+        }
+
+        window.localStorage.setItem(key, serialized);
+        return true;
+      } else {
+        this.memoryFallback.set(key, serialized);
+        return true;
+      }
+    } catch {
+      console.warn('[TraceLog] localStorage write failed');
+
+      if (this.available) {
+        this.cleanup();
+        try {
+          window.localStorage.setItem(key, JSON.stringify(value));
+        } catch {
+          this.memoryFallback.set(key, JSON.stringify(value));
+        }
+      } else {
+        this.memoryFallback.set(key, JSON.stringify(value));
+      }
+      return true;
+    }
+  }
+
+  get(key: string): unknown {
+    try {
+      let serialized: string | undefined = undefined;
+
+      if (this.available) {
+        serialized = window.localStorage.getItem(key) ?? undefined;
+      }
+
+      if (!serialized) {
+        serialized = this.memoryFallback.get(key) ?? undefined;
+      }
+
+      return serialized ? JSON.parse(serialized) : undefined;
+    } catch {
+      console.warn('[TraceLog] localStorage read failed');
+      return undefined;
+    }
+  }
+
+  remove(key: string): boolean {
+    try {
+      if (this.available) {
+        window.localStorage.removeItem(key);
+      }
+      this.memoryFallback.delete(key);
+      return true;
+    } catch {
+      console.warn('[TraceLog] localStorage remove failed');
+      return false;
+    }
+  }
+
+  isAvailable(): boolean {
+    return this.available;
+  }
+
+  clear(): void {
+    try {
+      if (this.available) {
+        const keys = Object.keys(window.localStorage);
+
+        for (const key of keys) {
+          if (key.startsWith('tracelog_') || key.includes('_critical_events') || key.includes('_heartbeat')) {
+            window.localStorage.removeItem(key);
+          }
+        }
+      }
+
+      this.memoryFallback.clear();
+    } catch {
+      console.warn('[TraceLog] localStorage clear failed');
+    }
+  }
+
+  getSize(): number {
+    try {
+      if (this.available) {
+        let size = 0;
+
+        for (const key in window.localStorage) {
+          if (key.startsWith('tracelog_') || key.includes('_critical_events') || key.includes('_heartbeat')) {
+            size += window.localStorage.getItem(key)?.length || 0;
+          }
+        }
+
+        return size;
+      }
+
+      let size = 0;
+
+      for (const value of this.memoryFallback.values()) {
+        size += value.length;
+      }
+
+      return size;
+    } catch {
+      return 0;
+    }
+  }
+
+  private cleanup(): void {
+    try {
+      if (!this.available) return;
+
+      const cutoffTime = Date.now() - 24 * 60 * 60 * 1000;
+      const keys = Object.keys(window.localStorage);
+
+      for (const key of keys) {
+        if (key.includes('_heartbeat') || key.includes('_critical_events')) {
+          try {
+            const data = JSON.parse(window.localStorage.getItem(key) || '{}');
+
+            if (data.timestamp && data.timestamp < cutoffTime) {
+              window.localStorage.removeItem(key);
+            }
+          } catch {
+            window.localStorage.removeItem(key);
+          }
+        }
+      }
+    } catch {
+      console.warn('[TraceLog] localStorage cleanup failed');
+    }
+  }
+}
